committing concurrent routines

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -7,7 +7,7 @@ from typing import List
 
 wait_random = __import__('0-basic_async_syntax').wait_random
 
-async def wait_n(n: int, max_delay: int) -> List[float]:
+async def wait_n(n: int, max_delay: int) -> List[int]:
     """Returns list of all delays in ascending order"""
     delays = [await wait_random(max_delay) for _ in range(n)]
     delays.sort()
